NXT-57 Add date filter label
