+++ web/src/components/GuidanceListPage/GuidanceListPageProps.ts
@@ -1,0 +1,12 @@
+import { SearchResults, SearchUrl } from "@nice-digital/search-client";
+
+export interface ActiveModifier {
+	displayName: string;
+	toggleUrl: string;
+}
+
+export interface GuidanceListPageProps {
+	results: SearchResults;
+	activeModifiers: ActiveModifier[];
+	searchUrl: SearchUrl;
+}
